Using local data
Having the CORS issue we cannot use the app to
get and save the data from the API, so i use local data instead.

The funcion 'likeSongLocalData' increase the likes when you press
it of the album.

The function 'commentSongLocalData' increase the comments only if
the input have any text.

Remove unnesary console.log.